Making a Player struct instead of an enum

+++ src/Commands/Modules/MoreGamesModule/MoreGamesModule.Multiplayer.cs
@@ -233,7 +233,7 @@ namespace PacManBot.Commands
                 await CancelGame();
                 return;
             }
-            if (!self && !user.IsBot && (game.UserId[(int)game.Turn] != user.Id || (DateTime.Now - game.LastPlayed) < TimeSpan.FromMinutes(1)))
+            if (!self && !user.IsBot && (game.UserId[game.Turn] != user.Id || (DateTime.Now - game.LastPlayed) < TimeSpan.FromMinutes(1)))
             {
                 await ReplyAsync("To remove another user they have to be inactive for at least 1 minute during their turn.");
             }
